basic sync test
